pyproject.toml
[project]
name = "tanda-suggester"
version = "0.1.0"
description = "Suggest tango tracks based on your own DJ history"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'",
    "pyqtdarktheme>=2.1",
    "pyside6>=6.6",
    "rapidfuzz>=3.0",
    "rich>=13.0",
]

[project.scripts]
tanda-suggester = "tanda_suggester.cli:cli"
tanda-suggester-gui = "tanda_suggester.gui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tanda_suggester"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pillow>=10.0",
    "pyinstaller>=6.0",
    "pytest>=9.0.3",
    "ruff>=0.15.9",
]
